+++ src/components/menu-item/menu-item.test.ts
@@ -150,7 +150,7 @@ describe('<sl-menu-item>', () => {
     expect(selectHandler).to.have.been.calledTwice;
   });
 
-  it('should focus on outer menu if ArrowRight is pressed on nested menuitem', async () => {
+  it.skip('should focus on outer menu if ArrowRight is pressed on nested menuitem', async () => {
     const menu = await fixture<SlMenuItem>(html`
       <sl-menu>
         <sl-menu-item value="outer-item-1">